+++ src/Network/UDP_ClientLinux.hpp
@@ -10,13 +10,13 @@
 namespace GLVM::core
 {
 	struct UDP_ClientLinux {
-		unsigned short port;
-		const char*    serverIP;
+		unsigned long port;
+		const char*   serverIP;
 		static const unsigned int  maxBufferSize = 1024;
-		int            socketFileDescriptor;
-		sockaddr_in    serverAddress;
-		const char*    message = "Hello, UDP server!";
-		char           buffer[maxBufferSize];
+		int           socketFileDescriptor;
+		sockaddr_in   serverAddress;
+		const char*   message = "Hello, UDP server!";
+		char          buffer[maxBufferSize];
 
 		UDP_ClientLinux( unsigned long port = 8080, const char* serverIP = "127.0.0.1" );
 		char* receive();
